Add gds-present command to wrap present
Add a wrapper for Go's present tool that sets the `-base` commandline
argument such that it points to the templates in this repository, passes
through any commandline arguments and pipes stdin, stderr and stdout
such that the wrapper is transparent to the user.

The `basePkg` constant must always match the import path for this
repository. I couldn't find a reliable way to detect the import path
automatically.