Add blue footer to each slide showing author name
Consistent with our current templates, add a blue footer showing the
presentation's authors and the GDS organisation name to all slides
except for the first and last.

Multiple authors are supported, consistent with present's file format.

Adding the footer required a number of changes to element padding to
allow the footer to span the full width of the slide.