Hide past and next slides
So that only one slide appears on the page at a time.

I haven't made any corresponding changes to the javascript as the
changes would be quite invasive, and this CSS changes has the desired
effect. This should make it easier for us to track upstream changes to
the Go present templates.